moved types to separate folder, updated import paths (#2)
