+++ src/components/Types/commentType.ts
@@ -1,0 +1,9 @@
+export type Comment = {
+    authorUuid: string
+    creationDate: number
+    description: string
+    rating: number
+    recipeUuid: string
+    type: string
+    uuid: string
+}
