+++ src/components/Types/profileType.ts
@@ -1,0 +1,7 @@
+export type ProfileType = {
+    username: string,
+    email: string,
+    picture: string,
+    description: string,
+    uuid: string
+}
